+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.9.6 - TBD
+ * Made `ExchangeWebServices::drillDownResponses()` static
+ * Added a `NotificationAPI::handlePushNotification()` method
+
 ## 0.9.6 - 2018-01-26
  * Added an `options` parameter to `CalendarAPI::updateCalendarItem()`